album-rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Image Knowledge Base RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi>=0.136.1",
    "openai>=2.33.0",
    "pillow>=12.2.0",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "rich>=15.0.0",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.46.0",
]

[project.scripts]
rag = "rag.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag"]

